@@ -10,18 +10,18 @@
     "marked": "^0.8.0",
     "moment": "^2.24.0",
     "polished": "^3.4.4",
-    "react": "^16.11.0",
+    "react": "^16.13.1",
     "react-dnd": "^9.5.1",
     "react-dnd-html5-backend": "^9.5.1",
-    "react-dom": "^16.11.0",
+    "react-dom": "^16.13.0",
     "react-feather": "^2.0.3",
     "react-form": "^4.0.1",
     "react-helmet": "^5.2.1",
     "react-modal": "^3.11.1",
     "react-query": "^1.1.0",
     "react-query-devtools": "^1.0.13",
-    "react-router": "^5.1.2",
-    "react-router-dom": "^5.1.2",
+    "react-router": "6.0.0-alpha.2",
+    "react-router-dom": "6.0.0-alpha.2",
     "react-select": "^3.1.0",
     "styled-components": "^5.0.0",
     "uuid": "^7.0.2"
