@@ -1,26 +1,28 @@
 import React, { useMemo } from 'react'
-import { useParams, useHistory } from 'react-router-dom'
+import { useParams, useNavigate } from 'react-router-dom'
 import { useForm } from 'react-form'
 
 import { AddIcon } from '../ui/icons'
 import { InputField } from './fields'
-import { mutateBoard } from '../resources/board'
+import { mutateBoard } from '../resources/boards'
 
-function AddBoardForm() {
-  const [mutate] = mutateBoard()
+function AddBoardForm({ name }) {
+  const { workspaceSlug,  boardSlug } = useParams()
+  const [mutate] = mutateBoard({ workspaceSlug, boardSlug })
 
-  const history = useHistory()
-  const { workspaceSlug } = useParams()
+  const navigate = useNavigate()
 
   const {
     Form,
     reset,
   } = useForm({
-    defaultValues: useMemo(() => ({ name: '' }), []),
+    defaultValues: useMemo(() => ({ name, }), [name]),
     onSubmit: async (values, instance) => {
       const { slug } = await mutate(values)
-      reset()
-      history.push(`/workspaces/${workspaceSlug}/${slug}`)
+      if (!boardSlug) {
+        reset()
+        navigate(`/workspaces/${workspaceSlug}/${slug}`)
+      }
     },
   })
 
@@ -29,6 +31,10 @@ function AddBoardForm() {
       <InputField icon={AddIcon} field="name" className="transparent full-width" placeholder="Add Board" />
     </Form>
   )
+}
+
+AddBoardForm.defaultProps = {
+  name: '',
 }
 
 export default AddBoardForm
