@@ -4,12 +4,31 @@ import moment from 'moment'
 
 import { useTicketVersions } from '../../resources'
 
+const CHANGES = [
+  ['board_id', ({ board_id }) => `moved ticket to board ${board_id.new}`],
+  ['column_id', ({ column_id }) => `moved ticket to column ${column_id.new}`],
+]
+
+const sentence = (by, changes) => {
+  const limit = CHANGES.length
+  let index = 0
+  let found = null
+  while(index < limit || found) {
+    const change = CHANGES[index]
+    if (Object.keys(changes).includes(change[0])) {
+      found = change[1](changes)
+      break
+    }
+    index += 1
+  }
+
+  return `${by.name} ${found || `updated ${Object.keys(changes).join(', ')}`}.`
+}
+
 const Version = ({ at, by, changes }) => {
   return (
     <div>
-      <div>
-        {by.name} updated {Object.keys(changes).join(', ')}
-      </div>
+      <div>{sentence(by, changes)}</div>
       <small>{moment(at).format('LLLL')}</small>
     </div>
   )
@@ -19,7 +38,7 @@ const Versions = ({ ticket }) => {
   const hasChanges = !moment(ticket.created_at).isSame(ticket.updated_at, 'seconds')
   const params = useParams()
   const { data: versions } = useTicketVersions(hasChanges && params)
-  // TODO initial data?
+  // TODO initial data? to avoid (versions || []) later
 
   return (
     <>
