@@ -1,28 +1,8 @@
 /* eslint-disable react-hooks/rules-of-hooks */
-import { useMyMutation } from './utils'
-import { useQuery, useMutation } from 'react-query'
+import { useQuery, useMutation, queryCache } from 'react-query'
 import api from './api'
 
-
-const getApiParameters = ({ workspaceSlug, boardSlug }) => pk => {
-  let method = 'post'
-  let path = ['', 'api', 'v1', 'workspaces', workspaceSlug]
-  if (boardSlug) {
-    path = path.concat(['boards', boardSlug])
-  }
-  path.push('tickets')
-  if (pk) {
-    path.push(pk)
-    method = 'patch'
-  }
-  path.push('')
-  return {
-    path: path.join('/'),
-    method,
-  }
-}
-
-const getTicketPath = ({ workspaceSlug, boardSlug, ticketSlug }) => {
+const getPath = ({ workspaceSlug, boardSlug, ticketSlug }) => {
   let method = 'post'
   let path = ['', 'workspaces', workspaceSlug]
   if (boardSlug) {
@@ -40,33 +20,34 @@ const getTicketPath = ({ workspaceSlug, boardSlug, ticketSlug }) => {
   }
 }
 
-const getTickets = async (key, params) => {
-  const { path } = getTicketPath(params)
+const get = async (key, params) => {
+  const { path } = getPath(params)
   const { data } = await api.get(path)
   return data
 }
 
 const getTicketVersions = async (key, params) => {
-  const { path } = getTicketPath(params)
+  const { path } = getPath(params)
   const { data } = await api.get(`${path}versions/`)
   return data
 }
 
-const createOrUpdateTicket = async (params, variables) => {
-  const { path, method, } = getTicketPath(params)
+const createOrUpdate = async (params, variables) => {
+  // TODO not clean
+  const { path, method, } = getPath({ ...params, ticketSlug: variables.pk || params.ticketSlug })
   const { data } = await api[method](path, variables)
   return data
 }
 
-const _deleteTicket = async (params) => {
-  const { path, } = getTicketPath(params)
+const _delete = async (params) => {
+  const { path, } = getPath(params)
   await api.delete(path)
   return {}
 }
 
 
 export const useTickets = (params) => {
-  return useQuery(['tickets', params], getTickets)
+  return useQuery(['tickets', params], get)
 }
 
 export const useTicketVersions = (params) => {
@@ -74,11 +55,15 @@ export const useTicketVersions = (params) => {
 }
 
 export const mutateTicket = (params) => {
-  const mutateFn = async data => await createOrUpdateTicket(params, data)
-  return useMutation(mutateFn)
+  const mutateFn = async data => await createOrUpdate(params, data)
+  return useMutation(mutateFn, {
+    onSuccess: (data, variables) => {
+      queryCache.refetchQueries(['tickets', params])
+    }
+  })
 }
 
 export const deleteTicket = (params) => {
-  const mutateFn = async data => await _deleteTicket(params)
+  const mutateFn = async data => await _delete(params)
   return useMutation(mutateFn)
 }
